crops/requirements.txt
PyYAML==5.3.1
markdown-table==2019.4.13